refactor: 알림 설정 컴포넌트를 BottomSheet에서 Dialog 방식으로 변경
