+++ app/src/main/java/com/conkeep/ui/feature/setting/SettingScreen.kt
@@ -20,7 +20,6 @@ import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
-import androidx.compose.material3.rememberModalBottomSheetState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.LaunchedEffect
@@ -258,10 +257,7 @@ fun SettingScreenContent(
     }
 
     showSettingBottomSheet.value?.let { initAlarmSetting: CouponAlarmSetting ->
-        AlarmSettingBottomSheet(
-            onDismiss = {
-                showSettingBottomSheet.value = null
-            },
+        AlarmSettingDialog(
             onConfirm = { selectedAlarm ->
                 addCouponAlarmSetting(selectedAlarm)
                 showSettingBottomSheet.value = null
@@ -270,7 +266,6 @@ fun SettingScreenContent(
             onCancel = {
                 showSettingBottomSheet.value = null
             },
-            sheetState = rememberModalBottomSheetState(skipPartiallyExpanded = true),
         )
     }
